feat: custom user agent InAppMessaging changes for UI handoff (#11639)
* chore: copy InAppMessaging to InternalInAppMessaging

* feat: add customUserAgentDetails param to InternalInAppMessaging

* feat: make InAppMessaging extend InternalInAppMessaging

* feat: send sync userAgentDetails through to service call

* fix: add SyncMessages value to InAppMessagingAction

* chore: remove extra parameter from non-service-call functions

* fix: fix typo in InAppMessagingAction

* feat: send user agent details through identify user to service call

* chore: remove unused import

* feat: allow InAppMessaging to be configured independently of Notifications

* test: update tests for internal changes

* build: export InternalInAppMessaging from scoped path

* fix: resolve bundle size

* chore: bundle size settings

* chore: increase bundle size limits

* chore: fix typo

* fix: update references to type with typo

* chore: update jsdocs in InternalInAppMessaging

* chore: increase bundle size limit

* chore: shorten jsdocs comment

* fix: first attempt to fix InAppMessaging internals

* fix: fix internals package.json paths

* test: fix import form internals

* chore: increase bundle size limits
