+++ packages/notifications/src/internals/index.ts
@@ -1,0 +1,4 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export { InternalInAppMessaging } from './InternalInAppMessaging';
