+++ packages/notifications/src/types.ts
@@ -22,7 +22,11 @@ export interface NotificationsProvider {
 	getProviderName(): string;
 
 	// identify the current user with the provider
-	identifyUser(userId: string, userInfo: UserInfo): Promise<void>;
+	identifyUser(
+		userId: string,
+		userInfo: UserInfo,
+		userAgentValue?: string
+	): Promise<void>;
 }
 
 export interface NotificationsConfig {
