+++ packages/predictions/package.json
@@ -63,7 +63,7 @@
 			"name": "Predictions (all providers)",
 			"path": "./lib-esm/index.js",
 			"import": "{ Amplify, Predictions, AmazonAIPredictionsProvider }",
-			"limit": "103.5 kB"
+			"limit": "103.55 kB"
 		},
 		{
 			"name": "Predictions (Convert provider)",
