+++ packages/pubsub/package.json
@@ -71,7 +71,7 @@
       "name": "PubSub (Mqtt provider)",
       "path": "./lib-esm/index.js",
       "import": "{ Amplify, PubSub, MqttOverWSProvider }",
-      "limit": "80.05 kB"
+      "limit": "80.1 kB"
     }
   ],
   "jest": {
